+++ packages/agtree/src/parser/index.ts
@@ -1,3 +1,4 @@
 export { RuleParser } from './rule-parser';
 export { defaultParserOptions, type ParserOptions } from './options';
 export { CosmeticRuleParser } from './cosmetic/cosmetic-rule-parser';
+export { DomainListParser } from './misc/domain-list-parser';